Hide internal types
